+++ CHANGELOG.md
@@ -29,6 +29,11 @@
 
 - OS2mo import job now use graphql API endpoint v7 and has configurable pagination size.
 
+### Bugfixes
+
+- Resolution timestamps are now correctly set whenever a result is automatically
+  handled.
+
 ## Version 3.20.0, 2nd June 2023
 
 "Quick Brown Fox"